Fix invalid usage of re.sub in wiki converter
The flags parameter was being passed in as the count parameter, so not all
links got converted properly (and probably caused some other issues as well).

Removed the flags parameter, since it was clearly not doing anything and
it's not even valid when used with a precompiled pattern.

Perhaps the compiled patterns need to be updated to include these flags? Will
add them later if it turns out to be needed.